[Refactor] small visibility changes in NodesCollector and DeploymentService for standalone
