+++ engine/interpreter/src/main/scala/pl/touk/nussknacker/engine/splittedgraph/SplittedNodesCollector.scala
@@ -1,0 +1,32 @@
+package pl.touk.nussknacker.engine.splittedgraph
+
+import pl.touk.nussknacker.engine.graph.node.NodeData
+import pl.touk.nussknacker.engine.splittedgraph.splittednode.{Case, EndingNode, FilterNode, Next, NextNode, OneOutputNode, PartRef, SplitNode, SplittedNode, SwitchNode}
+
+object SplittedNodesCollector {
+
+  def collectNodes(node: SplittedNode[_<:NodeData]): List[SplittedNode[_<:NodeData]] = {
+    val children = node match {
+      case n: OneOutputNode[_] =>
+        collectNodes(n.next)
+      case n: FilterNode =>
+        collectNodes(n.nextTrue) ::: n.nextFalse.toList.flatMap(collectNodes)
+      case n: SwitchNode =>
+        n.nexts.flatMap {
+          case Case(_, ch) => collectNodes(ch)
+        } ::: n.defaultNext.toList.flatMap(collectNodes)
+      case SplitNode(_, nextsWithParts) =>
+        nextsWithParts.flatMap(collectNodes)
+      case n: EndingNode[_] =>
+        List.empty
+    }
+    node :: children
+  }
+
+  private def collectNodes(next: Next): List[SplittedNode[_<:NodeData]] =
+    next match {
+      case NextNode(node) => collectNodes(node)
+      case part: PartRef => List.empty
+  }
+
+}
