+++ engine/standalone/app/src/main/scala/pl/touk/nussknacker/engine/standalone/http/ProcessRoute.scala
@@ -10,19 +10,19 @@ import cats.data.NonEmptyList
 import com.typesafe.scalalogging.LazyLogging
 import pl.touk.http.argonaut.Argonaut62Support
 import pl.touk.nussknacker.engine.api.exception.EspExceptionInfo
-import pl.touk.nussknacker.engine.standalone.StandaloneRequestHandler
-import pl.touk.nussknacker.engine.standalone.deployment.DeploymentService
+import pl.touk.nussknacker.engine.standalone.deployment.ProcessInterpreters
+import pl.touk.nussknacker.engine.standalone.{StandaloneProcessInterpreter, StandaloneRequestHandler}
 import pl.touk.nussknacker.engine.standalone.utils.logging.StandaloneRequestResponseLogger
 
 import scala.concurrent.ExecutionContext
 
-class ProcessRoute(deploymentService: DeploymentService) extends Directives with LazyLogging with Argonaut62Support {
+class ProcessRoute(processInterpreters: ProcessInterpreters) extends Directives with LazyLogging with Argonaut62Support {
 
   def route(log: StandaloneRequestResponseLogger)
            (implicit ec: ExecutionContext, mat: ActorMaterializer): Route =
     path(Segment) { processPath =>
       log.loggingDirective(processPath)(mat) {
-        deploymentService.getInterpreterByPath(processPath) match {
+        processInterpreters.getInterpreterByPath(processPath) match {
           case None =>
             complete {
               HttpResponse(status = StatusCodes.NotFound)
@@ -48,7 +48,7 @@ class ProcessRoute(deploymentService: DeploymentService) extends Directives with
     }
 
 
-  private def logErrors(processPath: JsonField, errors: NonEmptyList[EspExceptionInfo[_ <: Throwable]]) = {
+  private def logErrors(processPath: JsonField, errors: NonEmptyList[EspExceptionInfo[_ <: Throwable]]): Unit = {
     logger.warn(s"Failed to invoke: $processPath with errors: ${errors.map(_.throwable.getMessage)}")
     errors.toList.foreach { error =>
       logger.info(s"Invocation failed $processPath, error in ${error.nodeId}: ${error.throwable.getMessage}", error.throwable)
