+++ engine/standalone/engine/src/main/scala/pl/touk/nussknacker/engine/standalone/deployment/ProcessInterpreters.scala
@@ -1,0 +1,9 @@
+package pl.touk.nussknacker.engine.standalone.deployment
+
+import pl.touk.nussknacker.engine.standalone.StandaloneProcessInterpreter
+
+trait ProcessInterpreters {
+
+  def getInterpreterByPath(path: String): Option[StandaloneProcessInterpreter]
+
+}
